Opt for sync configure hook
